backend/requirements.txt
Flask_Cors==3.0.8
numpy==1.14.2
tulip_python==5.3.0
Flask==0.12.2
Werkzeug==0.14.1
pandas==0.23.4
umap-learn==0.3.10
